+++ dev/02-texts-for-translation.r
@@ -1,4 +1,5 @@
 library(dplyr)
+library(tidyr)
 library(purrr)
 library(writexl)
 
@@ -7,7 +8,7 @@ load_all()
 try(dir.create("dev/texts"))
 
 country_score %>%
-  distinct(country = country_territory) %>%
+  distinct(country) %>%
   bind_rows(
     country_rating_text %>%
       distinct(country)
@@ -32,21 +33,26 @@ country_score %>%
   distinct() %>%
   write_xlsx("dev/texts/continents.xlsx")
 
-country_score %>%
-  distinct(item_description) %>%
-  write_xlsx("dev/texts/score_items.xlsx")
+# country_score %>%
+#   distinct(item) %>%
+#   write_xlsx("dev/texts/items.xlsx")
 
 country_score %>%
+  distinct(item_description) %>%
+  write_xlsx("dev/texts/items_description.xlsx")
+
+country_rating_text %>%
   distinct(sub_item_description) %>%
-  write_xlsx("dev/texts/score_sub_items.xlsx")
+  drop_na() %>%
+  write_xlsx("dev/texts/sub_items_description.xlsx")
 
 country_rating_text %>%
   distinct(sub_item) %>%
-  write_xlsx("dev/texts/rating_sub_items.xlsx")
+  write_xlsx("dev/texts/sub_items.xlsx")
 
 country_rating_status %>%
   distinct(status) %>%
-  write_xlsx("dev/texts/rating_statuses.xlsx")
+  write_xlsx("dev/texts/statuses.xlsx")
 
 # country_rating_text %>%
 #   distinct(detail) %>%
@@ -54,8 +60,8 @@ country_rating_status %>%
 
 # save multiple files of 5,000 rows each
 detail <- country_rating_text %>%
-  distinct(detail)
+  select(detail)
 
 detail %>%
   split(1:nrow(detail) %/% 5000) %>%
-  walk2(seq_along(.), ~ write_xlsx(.x, paste0("dev/texts/rating_details_", .y, ".xlsx")))
+  walk2(seq_along(.), ~ write_xlsx(.x, paste0("dev/texts/details_", .y, ".xlsx")))