+++ NEWS.md
@@ -1,3 +1,8 @@
+# freedomhouse 0.3
+
+* Fixes some detail texts containing the sub item description, which is not
+  needed provided that there is a column for the sub item description
+  
 # freedomhouse 0.2
 
 * Adds 2023 data.